+++ tests/tests_pytorch/models/test_hooks.py
@@ -67,7 +67,7 @@ def test_on_before_zero_grad_called(tmpdir, max_steps):
     assert 0 == model.on_before_zero_grad_called
 
 
-def test_training_epoch_end_metrics_collection(tmpdir):
+def test_on_train_epoch_end_metrics_collection(tmpdir):
     """Test that progress bar metrics also get collected at the end of an epoch."""
     num_epochs = 3
 
@@ -77,7 +77,7 @@ def test_training_epoch_end_metrics_collection(tmpdir):
             self.log_dict({"step_metric": torch.tensor(-1), "shared_metric": 100}, logger=False, prog_bar=True)
             return output
 
-        def training_epoch_end(self, outputs):
+        def on_train_epoch_end(self):
             epoch = self.current_epoch
             # both scalar tensors and Python numbers are accepted
             self.log_dict(
@@ -97,40 +97,6 @@ def test_training_epoch_end_metrics_collection(tmpdir):
     # metrics are kept after each epoch
     for i in range(num_epochs):
         assert metrics[f"epoch_metric_{i}"] == i
-
-
-def test_training_epoch_end_metrics_collection_on_override(tmpdir):
-    """Test that batch end metrics are collected when training_epoch_end is overridden at the end of an epoch."""
-
-    class OverriddenModel(BoringModel):
-        def __init__(self):
-            super().__init__()
-            self.len_outputs = 0
-
-        def on_train_epoch_start(self):
-            self.num_train_batches = 0
-
-        def training_epoch_end(self, outputs):
-            self.len_outputs = len(outputs)
-
-        def on_train_batch_end(self, outputs, batch, batch_idx):
-            self.num_train_batches += 1
-
-    class NotOverriddenModel(BoringModel):
-        def on_train_epoch_start(self):
-            self.num_train_batches = 0
-
-        def on_train_batch_end(self, outputs, batch, batch_idx):
-            self.num_train_batches += 1
-
-    overridden_model = OverriddenModel()
-    not_overridden_model = NotOverriddenModel()
-    not_overridden_model.training_epoch_end = None
-
-    trainer = Trainer(max_epochs=1, default_root_dir=tmpdir, overfit_batches=2)
-
-    trainer.fit(overridden_model)
-    assert overridden_model.len_outputs == overridden_model.num_train_batches
 
 
 @pytest.mark.parametrize(
@@ -214,7 +180,6 @@ def test_transfer_batch_hook_ddp(tmpdir):
 
     model = TestModel()
     model.validation_step = None
-    model.training_epoch_end = None
     trainer = Trainer(
         default_root_dir=tmpdir,
         limit_train_batches=2,
@@ -285,14 +250,6 @@ class HookedModel(BoringModel):
             partial_h = partial(call, h, attr)
             update_wrapper(partial_h, attr)
             setattr(self, h, partial_h)
-
-    def validation_epoch_end(self, *args, **kwargs):
-        # `BoringModel` does not have a return for `validation_step_end` so this would fail
-        pass
-
-    def test_epoch_end(self, *args, **kwargs):
-        # `BoringModel` does not have a return for `test_step_end` so this would fail
-        pass
 
     def _train_batch(self, *args, **kwargs):
         if self.automatic_optimization:
@@ -391,12 +348,10 @@ class HookedModel(BoringModel):
 
     @staticmethod
     def _eval_epoch(fn, trainer, model, batches, key, device=torch.device("cpu")):
-        outputs = {key: ANY}
         return [
             dict(name=f"Callback.on_{fn}_epoch_start", args=(trainer, model)),
             dict(name=f"on_{fn}_epoch_start"),
             *HookedModel._eval_batch(fn, trainer, model, batches, key, device=device),
-            dict(name=f"{fn}_epoch_end", args=([outputs] * batches,)),
             dict(name=f"Callback.on_{fn}_epoch_end", args=(trainer, model)),
             dict(name=f"on_{fn}_epoch_end"),
         ]
@@ -546,7 +501,6 @@ def test_trainer_model_hook_system_fit(tmpdir, kwargs, automatic_optimization):
         dict(name="on_validation_end"),
         dict(name="train", args=(True,)),
         dict(name="on_validation_model_train"),
-        dict(name="training_epoch_end", args=([dict(loss=ANY)] * train_batches,)),
         dict(name="Callback.on_train_epoch_end", args=(trainer, model)),
         dict(name="on_train_epoch_end"),  # before ModelCheckpoint because it's a "monitoring callback"
         # `ModelCheckpoint.save_checkpoint` is called here
@@ -625,7 +579,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_epochs(tmpdir):
         dict(name="Callback.on_train_epoch_start", args=(trainer, model)),
         dict(name="on_train_epoch_start"),
         *model._train_batch(trainer, model, 2, current_epoch=1, current_batch=0),
-        dict(name="training_epoch_end", args=([dict(loss=ANY)] * 2,)),
         dict(name="Callback.on_train_epoch_end", args=(trainer, model)),
         dict(name="on_train_epoch_end"),  # before ModelCheckpoint because it's a "monitoring callback"
         # `ModelCheckpoint.save_checkpoint` is called here
@@ -705,7 +658,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_steps(tmpdir):
         dict(name="Callback.on_train_epoch_start", args=(trainer, model)),
         dict(name="on_train_epoch_start"),
         *model._train_batch(trainer, model, steps_after_reload, current_batch=1),
-        dict(name="training_epoch_end", args=([dict(loss=ANY)] * train_batches,)),
         dict(name="Callback.on_train_epoch_end", args=(trainer, model)),
         dict(name="on_train_epoch_end"),  # before ModelCheckpoint because it's a "monitoring callback"
         # `ModelCheckpoint.save_checkpoint` is called here
@@ -791,7 +743,6 @@ def test_trainer_model_hook_system_predict(tmpdir):
         dict(name="Callback.on_predict_epoch_start", args=(trainer, model)),
         dict(name="on_predict_epoch_start"),
         *model._predict_batch(trainer, model, batches),
-        # TODO: `predict_epoch_end`
         dict(name="Callback.on_predict_epoch_end", args=(trainer, model, [[ANY] * batches])),
         dict(name="on_predict_epoch_end", args=([[ANY] * batches],)),
         dict(name="Callback.on_predict_end", args=(trainer, model)),
@@ -837,7 +788,6 @@ def test_hooks_with_different_argument_names(tmpdir):
             return [DataLoader(RandomDataset(32, 64)), DataLoader(RandomDataset(32, 64))]
 
     model = CustomBoringModel()
-    model.test_epoch_end = None
 
     trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=5)